pyproject.toml
[project]
name = "learming"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Tobi DEGNON", email = "tobidegnon@proton.me" }
]
dependencies = [
    "Django",
    "django-click",
    "django-test-plus",
    "model-bakery",
    "pre-commit",
    "psycopg2-binary",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "whitenoise",
    "django-environ>=0.11.2",
    "gunicorn>=21.2.0",
    "django-q2>=1.5.5",
    "django-schema-viewer>=0.1.0",
    "django-allauth>=0.56.1",
    "django-model-utils>=4.3.1",
    "django-improved-user[factory]>=2.0a2",
    "django-htmx>=1.16.0",
    "django-pydantic-field>=0.2.10",
    "django-lifecycle>=1.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"
license = { text = "MIT" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
target-version = ["py310"]

[tool.coverage.run]
omit = [
    "*/admin.py",
    "*/manage.py",
    "*/migrations/*",
    "*/tests/*",
    "conftest.py",
]

[tool.pytest.ini_options]
addopts = "--cov --nomigrations --reuse-db"
DJANGO_SETTINGS_MODULE = "config.settings"
norecursedirs = ".git* frontend requirements templates"
python_files = "test_*.py"

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = ["E501", "E741"] # temporary

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".github",
    ".hg",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "migrations",
    "node_modules",
    "static",
]

per-file-ignores = {}

# Same as Black.
line-length = 120

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11.
target-version = "py311"
